refactor(organization_project): move to organization module
