+++ internal/sdkprovider/service/organization/organization_project_data_source.go
@@ -1,4 +1,4 @@
-package project
+package organization
 
 import (
 	"context"
